Update supabase interactions to new architecture
