+++ .vscode/launch.json
@@ -4,7 +4,6 @@
     // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
     "version": "0.2.0",
     "configurations": [
-        
         {
             "name": "Flutter",
             "request": "launch",
@@ -14,7 +13,7 @@
             "name": "Simulator",
             "request": "launch",
             "type": "dart",
-            "program": "front/lib/main.dart",
+            "program": "lib/main.dart",
             "args": [
                 "--flavor",
                 "dev",
