+++ front/lib/main.dart
@@ -1,19 +1,8 @@
-import 'package:flutter_dotenv/flutter_dotenv.dart';
-import 'package:supabase_flutter/supabase_flutter.dart';
+import 'package:front/screens/home/home_screen.dart';
+
 import 'commons.dart';
 
 Future<void> main() async {
-  WidgetsFlutterBinding.ensureInitialized();
-
-  // Load environment variables
-  await dotenv.load(fileName: ".env");
-
-  final supabaseUrl = dotenv.env['SUPABASE_URL'];
-  final supabaseAnonKey = dotenv.env['SUPABASE_ANON_KEY'];
-  if (supabaseUrl == null || supabaseAnonKey == null) {
-    throw StateError('Missing SUPABASE_URL or SUPABASE_ANON_KEY in .env');
-  }
-  await Supabase.initialize(url: supabaseUrl, anonKey: supabaseAnonKey);
   await initializeApp();
   runApp(const MyApp());
 }
@@ -29,41 +18,8 @@ class MyApp extends StatelessWidget {
       navigatorKey: navigatorKey,
 
       ///Screen names used from file screens.dart
-      initialRoute: Screens.profile,
-      routes: {Screens.profile: (_) => const ProfileScreen()},
-    );
-  }
-}
-
-class HomePage extends StatefulWidget {
-  const HomePage({super.key});
-
-  @override
-  State<HomePage> createState() => _HomePageState();
-}
-
-class _HomePageState extends State<HomePage> {
-  final _future = Supabase.instance.client.from('instruments').select();
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: FutureBuilder(
-        future: _future,
-        builder: (context, snapshot) {
-          if (!snapshot.hasData) {
-            return const Center(child: CircularProgressIndicator());
-          }
-          final instruments = snapshot.data!;
-          return ListView.builder(
-            itemCount: instruments.length,
-            itemBuilder: ((context, index) {
-              final instrument = instruments[index];
-              return ListTile(title: Text(instrument['name']));
-            }),
-          );
-        },
-      ),
+      initialRoute: Screens.home,
+      routes: {Screens.home: (_) => const HomeScreen()},
     );
   }
 }
