+++ front/lib/screens.dart
@@ -10,6 +10,5 @@ class Screens {
   static const welcome = '/welcome';
   static const login = '/login';
   static const profile = '/profile';
-
-  /// ...
+  static const home = '/home';
 }
