+++ front/lib/screens/home/home_controller.dart
@@ -1,0 +1,17 @@
+part of 'home_screen.dart';
+
+abstract class HomeScreenController extends State<HomeScreen> {
+  late final Future<List<dynamic>> _future = Supabase.instance.client
+      .from('instruments')
+      .select();
+
+  @override
+  void initState() {
+    super.initState();
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+  }
+}
