+++ front/lib/utils/utils.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/services.dart';
+import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:front/commons.dart';
+import 'package:supabase_flutter/supabase_flutter.dart';
 
 /// Used to initialize any services or other important things for the app
 Future<void> initializeApp() async {
@@ -17,4 +19,14 @@ Future<void> initializeApp() async {
       statusBarIconBrightness: Brightness.light,
     ),
   );
+
+  // Load environment variables
+  await dotenv.load(fileName: ".env");
+
+  final supabaseUrl = dotenv.env['SUPABASE_URL'];
+  final supabaseAnonKey = dotenv.env['SUPABASE_ANON_KEY'];
+  if (supabaseUrl == null || supabaseAnonKey == null) {
+    throw StateError('Missing SUPABASE_URL or SUPABASE_ANON_KEY in .env');
+  }
+  await Supabase.initialize(url: supabaseUrl, anonKey: supabaseAnonKey);
 }
